rename debug-log-file to reflect non-debug-only behaviour
