+++ packages/wrangler/src/logger.ts
@@ -4,7 +4,7 @@ import CLITable from "cli-table3";
 import { formatMessagesSync } from "esbuild";
 import { getEnvironmentVariableFactory } from "./environment-variables/factory";
 import { getSanitizeLogs } from "./environment-variables/misc-variables";
-import { appendToDebugLogFile } from "./utils/debug-log-file";
+import { appendToDebugLogFile } from "./utils/log-file";
 import type { Message } from "esbuild";
 export const LOGGER_LEVELS = {
 	none: -1,
